+++ .gitignore
@@ -1,1 +1,4 @@
+dist/
+node_modules/
 .DS_Store
+package-lock.json
