+++ public/search.html
@@ -35,7 +35,7 @@
             }
             .search-input {
                 width: 100%;
-                background-image: url(search.png);
+                background-image: url(images/search.png);
                 background-repeat: no-repeat;
                 padding-left: 28px;
                 border: 0;
@@ -82,5 +82,5 @@
                 <br />
             </div>
     </body>
-    <script src="search.js"></script>
+    <script src="js/search.js"></script>
 </html>
