+++ src/utils/index.ts
@@ -1,0 +1,7 @@
+
+export const chromeEventPromise = (chromeEvent: chrome.events.Event<any>): Promise<any> => {
+  return new Promise(resolve => {
+    // problem: this should fire multiple times
+    chromeEvent.addListener(resolve)
+  });
+};
